+++ src/Libraries/SmartStore.Core/Plugins/PluginManager.cs
@@ -33,8 +33,6 @@ namespace SmartStore.Core.Plugins
     /// </summary>
     public class PluginManager
     {
-        #region Fields
-
         private static readonly ReaderWriterLockSlim Locker = new ReaderWriterLockSlim();
         private static DirectoryInfo _shadowCopyFolder;
         //private static readonly string _installedPluginsFilePath = CommonHelper.MapPath("~/App_Data/InstalledPlugins.txt");
@@ -44,9 +42,13 @@ namespace SmartStore.Core.Plugins
 		private static readonly ConcurrentDictionary<string, PluginDescriptor> _referencedPlugins = new ConcurrentDictionary<string, PluginDescriptor>(StringComparer.OrdinalIgnoreCase);
         private static HashSet<Assembly> _inactiveAssemblies = new HashSet<Assembly>();
 
-        #endregion
-
-        #region Methods
+		/// <summary>
+		/// Returns the virtual path of the plugins folder relative to the application
+		/// </summary>
+		public static string PluginsLocation
+		{
+			get { return _pluginsPath; }
+		}
 
 		/// <summary> 
 		/// Returns a collection of all referenced plugin assemblies that have been shadow copied
@@ -462,10 +464,6 @@ namespace SmartStore.Core.Plugins
             return !_inactiveAssemblies.Contains(assembly);
         }
 
-        #endregion
-
-        #region Utilities
-
 		private static void SetPrivateEnvPath()
 		{
 			string dir = Environment.Is64BitProcess ? "amd64" : "x86";
@@ -646,7 +644,5 @@ namespace SmartStore.Core.Plugins
             if (!folder.Parent.Name.Equals("Plugins", StringComparison.InvariantCultureIgnoreCase)) return false;
             return true;
         }
-
-        #endregion
     }
 }