+++ src/Presentation/SmartStore.Web.Framework/Theming/ThemeVarsRepository.cs
@@ -3,11 +3,10 @@ using System.Dynamic;
 using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
-using SmartStore.Core.Caching;
 using SmartStore.Core.Infrastructure;
 using SmartStore.Services.Themes;
 using System.Web;
-using System.Web.Caching;
+using System.IO;
 
 namespace SmartStore.Web.Framework.Theming
 { 
@@ -16,17 +15,19 @@ namespace SmartStore.Web.Framework.Theming
         private static readonly Regex s_keyWhitelist = new Regex(@"^[a-zA-Z0-9_-]+$");
         private static readonly Regex s_valueWhitelist = new Regex(@"^[#@]?[a-zA-Z0-9""' _\.,-]*$");
 
-        public string GetVariablesAsLess(string themeName, int storeId)
+		public string GetPreprocessorCss(string extension, string themeName, int storeId)
         {
             Guard.NotEmpty(themeName, nameof(themeName));
             Guard.IsPositive(storeId, nameof(storeId));
 
-            var variables = GetLessCssVariables(themeName, storeId);
-            var lessCss = TransformToLess(variables);
-            return lessCss;
+            var variables = GetVariables(themeName, storeId);
+
+			var isLess = extension.IsCaseInsensitiveEqual(".less");
+            var css = Transform(variables, isLess);
+            return css;
         }
 
-        private IDictionary<string, string> GetLessCssVariables(string themeName, int storeId)
+        private IDictionary<string, string> GetVariables(string themeName, int storeId)
         {
             var result = new Dictionary<string, string>();
 
@@ -52,7 +53,7 @@ namespace SmartStore.Web.Framework.Theming
 
         internal virtual ExpandoObject GetRawVariables(string themeName, int storeId)
         {
-			// we need the Asp.Net here cache in order to define cacheKey dependencies
+			// we need the Asp.Net cache here in order to define cacheKey dependencies
 
 			string cacheKey = FrameworkCacheConsumer.BuildThemeVarsCacheKey(themeName, storeId);
 
@@ -63,20 +64,21 @@ namespace SmartStore.Web.Framework.Theming
 			});
         }
 
-        private string TransformToLess(IDictionary<string, string> parameters)
-        {
-            if (parameters.Count == 0)
-                return string.Empty;
+		private string Transform(IDictionary<string, string> parameters, bool toLess)
+		{
+			if (parameters.Count == 0)
+				return string.Empty;
 
-            var sb = new StringBuilder();
-            foreach (var parameter in parameters.Where(kvp => kvp.Value.HasValue()))
-            {
-                sb.AppendFormat("@{0}: {1};\n", parameter.Key, parameter.Value);
-            }
+			var prefix = toLess ? "@" : "$";
 
-            return sb.ToString();
-        }
+			var sb = new StringBuilder();
+			foreach (var parameter in parameters.Where(kvp => kvp.Value.HasValue()))
+			{
+				sb.AppendFormat("{0}{1}: {2};\n", prefix, parameter.Key, parameter.Value);
+			}
 
-    }
+			return sb.ToString();
+		}
+	}
 
 }